Make coverage drop check async
